zero-shot/pyproject.toml
[tool.poetry]
name = "masterthesis"
version = "0.1.0"
description = ""
readme = "README.md"


[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest<9.0.0,>=7.0.0",
    "pre-commit<4.0.0,>=3.7.1",
    "sphinx<8.0.0,>=7.3.7",
    "prospector[with-bandit,with-mypy]<2.0.0,>=1.10.3",
    "syrupy<5.0.0,>=4.6.1",
    "ipykernel<7.0.0,>=6.29.4",
    "docker<8.0.0,>=7.1.0",
]

[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[project]

requires-python = "<3.13,>=3.12"
dependencies = [
    "groq>=0.31.1",
    "graphviz<1.0.0,>=0.20.3",
    "pylspclient>=0.0.3",
    "ipykernel>=6.29.4",
    "docker>=7.1.0",
    "pytest-mock>=3.14.0",
    "syrupy>=4.6.1",
    "patch==1.*",
    "diff-match-patch>=20230430",
    "python-dotenv>=1.0.1",
    "mistletoe>=1.3.0",
    "tiktoken>=0.4.0",
    "GitPython>=3.1.43",
    "matplotlib>=3.9.1",
    "seaborn>=0.13.2",
    "langchain>=0.2.5",
    "langchain-openai>=0.1.14",
    "arize-phoenix>=4.7.1",
    "openinference-instrumentation-dspy>=0.1.9",
    "opentelemetry-exporter-otlp>=1.25.0",
    "black[jupyter]>=24.4.2",
    "transformers>=4.42.3",
    "torch>=2.3.1",
    "pathvalidate>=3.2.0",
    "json-schema-codegen>=0.6.3",
    "mypy>=1.10.1",
    "optuna>=3.6.1",
    "optuna-dashboard>=0.15.1",
    "anthropic>=0.31.2",
    "mistral-common>=1.3.3",
    "plotly>=5.23.0",
    "nbformat>=5.10.4",
    "isort>=5.13.2",
    "autoflake>=1.7.8",
    "PyYAML>=6.0.1",
    "mistralai>=0.4.2",
    "jsonschema>=4.21.1",
    "check-jsonschema>=0.29.0",
    "vertexai==1.43.0",
    "google-cloud-aiplatform==1.43.0",
    "dspy[vertex-anthropic] @ git+https://github.com/LukvonStrom/dspy.git@vertex-anthropic",
    "ijson>=3.3.0",
]
name = "masterthesis"
version = "0.1.0"
description = ""
readme = "README.md"

[project.optional-dependencies]
analysis = [
    "ipywidgets>=8.1.3",
    "kaleido==0.2.1",
    "nbconvert>=7.16.4",
    "nbstripout>=0.7.1",
    "unidiff>=0.7.5",
]
[tool.black]
extend-exclude = '''/dataset'''